build(developers): updates links and page names
