+++ src/pages/index.js
@@ -1,40 +1,32 @@
 import clsx from 'clsx';
-import Link from '@docusaurus/Link';
+//import Link from '@docusaurus/Link';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import Layout from '@theme/Layout';
-import HomepageFeatures from '@site/src/components/HomepageFeatures';
-import HeroImageUrl from '@site/static/img/hero.webp';
-import Translate from '@docusaurus/Translate';
-import Heading from '@theme/Heading';
+import DevelopersHomepage from '@site/src/components/DevelopersHomepage';
+import HeroImageUrl from '@site/static/img/developers/hero.png';
+//import Translate from '@docusaurus/Translate';
+//import Heading from '@theme/Heading';
 import styles from './index.module.css';
+import React from 'react';
+import { Banner, Button, Typography } from '@teradata-web/react-components';
+
+
 function HomepageHeader() {
   const { siteConfig } = useDocusaurusContext();
   return (
     <header className={clsx('hero--primary', styles.heroBanner)}>
       <div className={clsx('container', styles.container)}>
-        <div className="">
-          <Heading as="h1" className="hero__title">
-            {siteConfig.title}
-          </Heading>
-          <p className="hero__subtitle">
-            <Translate id="theme.tagline">{siteConfig.tagline}</Translate>
-          </p>
-          <div className={styles.buttons}>
-            <Link
-              className={clsx(
-                'button button--secondary button--lg',
-                styles.getStartedButton
-              )}
-              to="/docs/install-ai-unlimited"
-            >
-              <Translate id="theme.button">Get Started</Translate>
-            </Link>
-          </div>
-        </div>
-        <div className={styles.heroImage}>
-          <img src={HeroImageUrl} width="498" alt="A woman smiling and holding a laptop, standing in a modern office environment with abstract geometric shapes in the background." />
-          <div className={styles.heroBlur}></div>
-        </div>
+        <Banner
+            content={
+              <div>
+                <Typography scale="body1">A scalable, on-demand compute engine in the cloud.</Typography> 
+                <br/>
+                <Button label='Learn more' primary={true} icon='fa fa-arrow-right' trailingIcon={true}></Button>
+              </div>}
+            imageSrc={HeroImageUrl}
+            isHero={true}
+            title="Welcome to the Teradata Developers Portal"
+          />
       </div>
     </header>
   );
@@ -43,12 +35,10 @@ function HomepageHeader() {
 export default function Home() {
   const { siteConfig } = useDocusaurusContext();
   return (
-    <Layout
-      description={siteConfig.tagline}
-    >
+    <Layout description={siteConfig.tagline}>
       <HomepageHeader />
       <main className={styles.features}>
-        <HomepageFeatures />
+        <DevelopersHomepage />
       </main>
     </Layout>
   );
